+++ subprojects/common/analysis/src/test/java/hu/bme/mit/theta/analysis/algorithm/BoundedTest.kt
@@ -28,7 +28,7 @@ import hu.bme.mit.theta.core.type.inttype.IntExprs
 import hu.bme.mit.theta.core.type.inttype.IntExprs.Int
 import hu.bme.mit.theta.core.utils.StmtUtils
 import hu.bme.mit.theta.core.utils.indexings.VarIndexingFactory
-import hu.bme.mit.theta.solver.z3.Z3LegacySolverFactory
+import hu.bme.mit.theta.solver.z3legacy.Z3LegacySolverFactory
 import org.junit.Assert
 import org.junit.Test
