+++ subprojects/common/analysis/src/test/java/hu/bme/mit/theta/analysis/expl/ExplInitFuncTest.java
@@ -32,7 +32,7 @@ import hu.bme.mit.theta.common.Utils;
 import hu.bme.mit.theta.core.decl.VarDecl;
 import hu.bme.mit.theta.core.type.inttype.IntType;
 import hu.bme.mit.theta.solver.Solver;
-import hu.bme.mit.theta.solver.z3.Z3LegacySolverFactory;
+import hu.bme.mit.theta.solver.z3legacy.Z3LegacySolverFactory;
 
 public class ExplInitFuncTest {
