+++ subprojects/common/analysis/src/test/java/hu/bme/mit/theta/analysis/expl/ExplStatePredicateTest.java
@@ -38,7 +38,7 @@ import hu.bme.mit.theta.core.type.Expr;
 import hu.bme.mit.theta.core.type.booltype.BoolType;
 import hu.bme.mit.theta.core.type.inttype.IntType;
 import hu.bme.mit.theta.solver.Solver;
-import hu.bme.mit.theta.solver.z3.Z3LegacySolverFactory;
+import hu.bme.mit.theta.solver.z3legacy.Z3LegacySolverFactory;
 
 @RunWith(Parameterized.class)
 public class ExplStatePredicateTest {
