+++ subprojects/common/analysis/src/test/java/hu/bme/mit/theta/analysis/expr/ExprTraceCheckersTest.java
@@ -28,7 +28,7 @@ import hu.bme.mit.theta.core.type.inttype.IntType;
 import hu.bme.mit.theta.core.utils.indexings.VarIndexingFactory;
 import hu.bme.mit.theta.solver.ItpSolver;
 import hu.bme.mit.theta.solver.UCSolver;
-import hu.bme.mit.theta.solver.z3.Z3LegacySolverFactory;
+import hu.bme.mit.theta.solver.z3legacy.Z3LegacySolverFactory;
 import org.junit.Before;
 import org.junit.Test;
