+++ subprojects/common/analysis/src/test/java/hu/bme/mit/theta/analysis/pred/PredOrdTest.java
@@ -26,7 +26,7 @@ import org.junit.Test;
 import hu.bme.mit.theta.core.decl.Decls;
 import hu.bme.mit.theta.core.decl.VarDecl;
 import hu.bme.mit.theta.core.type.inttype.IntType;
-import hu.bme.mit.theta.solver.z3.Z3LegacySolverFactory;
+import hu.bme.mit.theta.solver.z3legacy.Z3LegacySolverFactory;
 
 public class PredOrdTest {
