+++ subprojects/solver/solver-z3-legacy/src/main/java/hu/bme/mit/theta/solver/z3legacy/Z3ItpMarker.java
@@ -13,31 +13,38 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package hu.bme.mit.theta.solver.z3;
+package hu.bme.mit.theta.solver.z3legacy;
+
+import hu.bme.mit.theta.solver.ItpMarker;
+import hu.bme.mit.theta.solver.Stack;
+import hu.bme.mit.theta.solver.impl.StackImpl;
+
+import java.util.Collection;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
-import java.util.Map;
+final class Z3ItpMarker implements ItpMarker {
 
-import hu.bme.mit.theta.core.type.Expr;
-import hu.bme.mit.theta.core.type.booltype.BoolType;
-import hu.bme.mit.theta.solver.Interpolant;
-import hu.bme.mit.theta.solver.ItpMarker;
+    private final Stack<com.microsoft.z3.BoolExpr> terms;
 
-final class Z3Interpolant implements Interpolant {
-
-    private final Map<ItpMarker, Expr<BoolType>> itpMap;
-
-    Z3Interpolant(final Map<ItpMarker, Expr<BoolType>> itpMap) {
-        this.itpMap = itpMap;
+    public Z3ItpMarker() {
+        terms = new StackImpl<>();
     }
 
-    @Override
-    public Expr<BoolType> eval(final ItpMarker marker) {
-        checkNotNull(marker);
-        final Expr<BoolType> itpExpr = itpMap.get(marker);
-        checkNotNull(itpExpr);
-        return itpExpr;
+    public void add(final com.microsoft.z3.BoolExpr term) {
+        terms.add(checkNotNull(term));
+    }
+
+    public void push() {
+        terms.push();
+    }
+
+    public void pop(final int n) {
+        terms.pop(n);
+    }
+
+    public Collection<com.microsoft.z3.BoolExpr> getTerms() {
+        return terms.toCollection();
     }
 
 }
