+++ subprojects/solver/solver-z3-legacy/src/test/java/hu/bme/mit/theta/solver/z3/Z3ModelTest.java
@@ -19,6 +19,7 @@ import com.microsoft.z3.BoolExpr;
 import com.microsoft.z3.Context;
 import com.microsoft.z3.Model;
 import com.microsoft.z3.Solver;
+import hu.bme.mit.theta.solver.z3legacy.Z3LegacySolverFactory;
 import org.junit.Assert;
 import org.junit.Test;
