+++ subprojects/solver/solver-z3-legacy/src/test/java/hu/bme/mit/theta/solver/z3/Z3SolverTest.java
@@ -36,6 +36,7 @@ import hu.bme.mit.theta.core.type.inttype.IntType;
 import hu.bme.mit.theta.solver.Solver;
 import hu.bme.mit.theta.solver.SolverStatus;
 import hu.bme.mit.theta.solver.UCSolver;
+import hu.bme.mit.theta.solver.z3legacy.Z3LegacySolverFactory;
 import org.junit.Before;
 import org.junit.Test;
